ref(slack): Simplify assistant thread delivery
Move assistant-thread lifecycle, status, title generation, and mrkdwn translation into Slack-owned modules. Remove the old progress reporter and transport split so the outbound contract is smaller and Slack-specific behavior stays isolated.

Keep the existing assistant status and title semantics while simplifying the reply planning path and tightening the integration coverage around the Chat SDK and Slack API wiring.

Refs GH-202
