+++ packages/junior/src/chat/slack/assistant-thread/lifecycle.ts
