+++ packages/junior/src/chat/slack/mrkdwn.ts
@@ -1,0 +1,71 @@
+import { truncateStatusText } from "@/chat/runtime/status-format";
+
+/** Insert blank lines between content blocks so Slack renders them with visual separation. */
+export function ensureBlockSpacing(text: string): string {
+  const codeBlockPattern = /^```/;
+  const listItemPattern = /^[-*•]\s|^\d+\.\s/;
+  const lines = text.split("\n");
+  const result: string[] = [];
+  let inCodeBlock = false;
+
+  for (let i = 0; i < lines.length; i++) {
+    const line = lines[i];
+    const isCodeFence = codeBlockPattern.test(line.trimStart());
+
+    if (isCodeFence) {
+      if (!inCodeBlock) {
+        const prev = result.length > 0 ? result[result.length - 1] : undefined;
+        if (prev !== undefined && prev.trim() !== "") {
+          result.push("");
+        }
+      }
+      inCodeBlock = !inCodeBlock;
+      result.push(line);
+      continue;
+    }
+
+    if (inCodeBlock) {
+      result.push(line);
+      continue;
+    }
+
+    const prev = result.length > 0 ? result[result.length - 1] : undefined;
+    if (
+      prev !== undefined &&
+      prev.trim() !== "" &&
+      line.trim() !== "" &&
+      !(
+        listItemPattern.test(prev.trimStart()) &&
+        listItemPattern.test(line.trimStart())
+      )
+    ) {
+      result.push("");
+    }
+
+    result.push(line);
+  }
+
+  return result.join("\n");
+}
+
+/**
+ * Render model-authored markdown into Slack-friendly `mrkdwn`.
+ *
+ * Slack reply delivery owns chunking and continuation markers separately.
+ * This helper only normalizes text into the repository's canonical Slack
+ * rendering form.
+ */
+export function renderSlackMrkdwn(text: string): string {
+  let normalized = text.replace(/\r\n?/g, "\n").replace(/[ \t]+$/gm, "");
+  normalized = ensureBlockSpacing(normalized);
+  return normalized.replace(/\n{3,}/g, "\n\n").trim();
+}
+
+/** Normalize a typed assistant status context before handing it to Slack. */
+export function normalizeSlackStatusText(text: string): string {
+  const trimmed = text.trim();
+  if (!trimmed) {
+    return "";
+  }
+  return truncateStatusText(trimmed.replace(/(?:\.\s*)+$/, "").trim());
+}
