+++ packages/junior/src/chat/slack/reply.ts
@@ -1,7 +1,7 @@
 import { Buffer } from "node:buffer";
-import type { FileUpload, PostableMessage } from "chat";
+import type { FileUpload } from "chat";
 import type { AssistantReply } from "@/chat/respond";
-import { resolveReplyDelivery } from "@/chat/services/reply-delivery-plan";
+import type { ReplyFileDelivery } from "@/chat/services/reply-delivery-plan";
 import { uploadFilesToThread } from "@/chat/slack/outbound";
 import {
   buildSlackOutputMessage,
@@ -14,22 +14,52 @@ export type PlannedSlackReplyStage =
   | "thread_reply_files_followup";
 
 export interface PlannedSlackReplyPost {
-  message: PostableMessage;
+  files?: FileUpload[];
   stage: PlannedSlackReplyStage;
+  text: string;
 }
 
 function isInterruptedVisibleReply(reply: AssistantReply): boolean {
   return reply.diagnostics.outcome === "provider_error";
 }
 
-function buildChunkMessage(
-  chunk: string,
-  files?: FileUpload[],
-): PostableMessage {
-  return {
-    markdown: chunk,
-    ...(files ? { files } : {}),
+function resolveReplyDelivery(reply: AssistantReply): {
+  shouldPostThreadReply: boolean;
+  attachFiles: ReplyFileDelivery;
+} {
+  const replyHasFiles = Boolean(reply.files && reply.files.length > 0);
+  const deliveryPlan = reply.deliveryPlan ?? {
+    mode: reply.deliveryMode ?? "thread",
+    postThreadText: (reply.deliveryMode ?? "thread") !== "channel_only",
+    attachFiles: replyHasFiles ? "inline" : "none",
   };
+
+  return {
+    shouldPostThreadReply: deliveryPlan.postThreadText,
+    attachFiles:
+      replyHasFiles && deliveryPlan.attachFiles !== "none" ? "inline" : "none",
+  };
+}
+
+function buildReplyText(text: string): string {
+  const message = buildSlackOutputMessage(text);
+  if (
+    typeof message === "object" &&
+    message !== null &&
+    "markdown" in message &&
+    typeof message.markdown === "string"
+  ) {
+    return message.markdown;
+  }
+  if (
+    typeof message === "object" &&
+    message !== null &&
+    "raw" in message &&
+    typeof message.raw === "string"
+  ) {
+    return message.raw;
+  }
+  return "";
 }
 
 function buildTextPosts(args: {
@@ -42,10 +72,8 @@ function buildTextPosts(args: {
     interrupted: args.interrupted,
   });
   return chunks.map((chunk, index) => ({
-    message: buildChunkMessage(
-      chunk,
-      index === 0 ? args.firstFiles : undefined,
-    ),
+    text: chunk,
+    ...(index === 0 && args.firstFiles ? { files: args.firstFiles } : {}),
     stage:
       index === 0
         ? (args.firstStage ?? "thread_reply")
@@ -90,37 +118,6 @@ async function uploadReplyFilesBestEffort(args: {
   }
 }
 
-function getReplyMessageText(message: PostableMessage): string | undefined {
-  if (typeof message !== "object" || message === null) {
-    return undefined;
-  }
-
-  if ("markdown" in message && typeof message.markdown === "string") {
-    return message.markdown;
-  }
-
-  if ("raw" in message && typeof message.raw === "string") {
-    return message.raw;
-  }
-
-  return undefined;
-}
-
-function getReplyMessageFiles(
-  message: PostableMessage,
-): FileUpload[] | undefined {
-  if (
-    typeof message === "object" &&
-    message !== null &&
-    "files" in message &&
-    Array.isArray(message.files)
-  ) {
-    return message.files;
-  }
-
-  return undefined;
-}
-
 /**
  * Plan the Slack thread posts needed to realize a completed assistant reply,
  * including chunking, interruption markers, and file delivery.
@@ -132,9 +129,9 @@ export function planSlackReplyPosts(args: {
     args.reply.files && args.reply.files.length > 0
       ? args.reply.files
       : undefined;
-  const { shouldPostThreadReply, attachFiles } = resolveReplyDelivery({
-    reply: args.reply,
-  });
+  const { shouldPostThreadReply, attachFiles } = resolveReplyDelivery(
+    args.reply,
+  );
   const interrupted = isInterruptedVisibleReply(args.reply);
   const posts: PlannedSlackReplyPost[] = [];
 
@@ -149,20 +146,22 @@ export function planSlackReplyPosts(args: {
 
   if (attachFiles === "inline" && replyFiles && textPosts.length === 0) {
     posts.push({
-      message: buildSlackOutputMessage("", replyFiles),
+      files: replyFiles,
       stage: "thread_reply",
+      text: "",
     });
   } else if (shouldPostThreadReply && textPosts.length === 0) {
     posts.push({
-      message: buildSlackOutputMessage(args.reply.text),
+      text: buildReplyText(args.reply.text),
       stage: "thread_reply",
     });
   }
 
   if (attachFiles === "followup" && replyFiles) {
     posts.push({
-      message: buildSlackOutputMessage("", replyFiles),
+      files: replyFiles,
       stage: "thread_reply_files_followup",
+      text: "",
     });
   }
 
@@ -184,20 +183,18 @@ export async function postSlackApiReplyPosts(args: {
   ) => Promise<void>;
 }): Promise<void> {
   for (const post of args.posts) {
-    const text = getReplyMessageText(post.message);
-    if (text && text.trim().length > 0) {
-      await args.postMessage(args.channelId, args.threadTs, text);
+    if (post.text.trim().length > 0) {
+      await args.postMessage(args.channelId, args.threadTs, post.text);
     }
 
-    const files = getReplyMessageFiles(post.message);
-    if (!files?.length) {
+    if (!post.files?.length) {
       continue;
     }
 
     await uploadReplyFilesBestEffort({
       channelId: args.channelId,
       threadTs: args.threadTs,
-      files,
+      files: post.files,
     });
   }
 }
